+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export { default as ChoicePage } from '@/pages/Choice';
+export { default as ResultPage } from '@/pages/Result';
